# openapi.json
{
  "components": {
    "schemas": {
...
      "AutomationClientCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "client_id": {
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
            "title": "Client Id",
            "type": "string"
          },
          "name": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "redirect_uris": {
            "items": {
              "type": "string"
            },
            "maxItems": 10,
            "minItems": 1,
            "title": "Redirect Uris",
            "type": "array"
          },
          "scopes": {
            "items": {
              "enum": [
                "app.read",
                "app.write",
                "app.run",
                "app.cancel",
                "app.credentials.read",
                "app.credentials.write",
                "manager.read",
                "manager.runtime",
                "manager.mutate",
                "app.admin"
              ],
              "type": "string"
            },
            "minItems": 1,
            "title": "Scopes",
            "type": "array"
          }
        },
        "required": [
          "client_id",
          "name",
          "redirect_uris",
          "scopes"
        ],
        "title": "AutomationClientCreateRequest",
        "type": "object"
      },
...
      "BundleImportRequest": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Name"
          },
          "source_artifact_id": {
            "title": "Source Artifact Id",
            "type": "string"
          }
        },
        "required": [
          "source_artifact_id"
        ],
        "title": "BundleImportRequest",
        "type": "object"
      },
...
      "JobCreate": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Kind",
            "type": "string"
          },
          "max_attempts": {
            "default": 1,
            "maximum": 10,
            "minimum": 1,
            "title": "Max Attempts",
            "type": "integer"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Session Id"
          }
        },
        "required": [
          "kind"
        ],
        "title": "JobCreate",
        "type": "object"
      },
...
      "ManagerDesiredComponentState": {
        "additionalProperties": false,
        "properties": {
          "compute": {
            "default": "auto",
            "enum": [
              "auto",
              "cpu",
              "cuda",
              "vulkan",
              "metal",
              "rocm",
              "wgpu"
            ],
            "title": "Compute",
            "type": "string"
          },
          "options": {
            "additionalProperties": true,
            "title": "Options",
            "type": "object"
          },
          "present": {
            "default": true,
            "title": "Present",
            "type": "boolean"
          },
          "quantization": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Quantization"
          }
        },
        "title": "ManagerDesiredComponentState",
        "type": "object"
      },
...
      "ManagerPlanRequest": {
        "additionalProperties": false,
        "properties": {
          "desired": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ManagerDesiredComponentState"
            },
            "title": "Desired",
            "type": "object"
          },
          "expected_revision": {
            "anyOf": [
              {
                "minimum": 0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Expected Revision"
          },
          "kind": {
            "enum": [
              "install",
              "update",
              "repair",
              "remove",
              "uninstall",
              "start",
              "stop",
              "restart",
              "import"
            ],
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "desired"
        ],
        "title": "ManagerPlanRequest",
        "type": "object"
      },
...
      "ProviderCreate": {
        "additionalProperties": false,
        "properties": {
          "api_key": {
            "anyOf": [
              {
                "maxLength": 65536,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Api Key"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Base Url"
          },
          "credential_backend": {
            "anyOf": [
              {
                "enum": [
                  "database",
                  "environment",
                  "keyring",
                  "file"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Credential Backend"
          },
          "credential_reference": {
            "anyOf": [
              {
                "maxLength": 4096,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Credential Reference"
          },
          "delete_previous_credential": {
            "default": false,
            "title": "Delete Previous Credential",
            "type": "boolean"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "kind": {
            "default": "llm",
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Label",
            "type": "string"
          },
          "options": {
            "additionalProperties": true,
            "title": "Options",
            "type": "object"
          },
          "provider_key": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Provider Key",
            "type": "string"
          },
          "secret_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Secret Ref"
          }
        },
        "required": [
          "provider_key",
          "label"
        ],
        "title": "ProviderCreate",
        "type": "object"
      },
...
      "RvcConvertRequest": {
        "additionalProperties": false,
        "properties": {
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Session Id"
          },
          "settings": {
            "additionalProperties": true,
            "title": "Settings",
            "type": "object"
          },
          "source_artifact_id": {
            "title": "Source Artifact Id",
            "type": "string"
          }
        },
        "required": [
          "source_artifact_id"
        ],
        "title": "RvcConvertRequest",
        "type": "object"
      },
...
      "TokenCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "expires_in_days": {
            "anyOf": [
              {
                "maximum": 365,
                "minimum": 1,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Expires In Days"
          },
          "label": {
            "default": "CLI token",
            "maxLength": 160,
            "title": "Label",
            "type": "string"
          },
          "scopes": {
            "items": {
              "enum": [
                "app.read",
                "app.write",
                "app.run",
                "app.cancel",
                "app.credentials.read",
                "app.credentials.write",
                "manager.read",
                "manager.runtime",
                "manager.mutate",
                "app.admin"
              ],
              "type": "string"
            },
            "minItems": 1,
            "title": "Scopes",
            "type": "array"
          }
        },
        "required": [
          "scopes"
        ],
        "title": "TokenCreateRequest",
        "type": "object"
      },
      "TrainingCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "model_name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Model Name",
            "type": "string"
          },
          "settings": {
            "additionalProperties": true,
            "title": "Settings",
            "type": "object"
          },
          "source_artifact_id": {
            "title": "Source Artifact Id",
            "type": "string"
          },
          "source_text_artifact_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Source Text Artifact Id"
          },
          "voice_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Voice Id"
          }
        },
        "required": [
          "model_name",
          "source_artifact_id"
        ],
        "title": "TrainingCreateRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerToken": {
        "scheme": "bearer",
        "type": "http"
      },
      "cookieAuth": {
        "in": "cookie",
        "name": "session",
        "type": "apiKey"
      },
      "nativeOAuth": {
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "/api/v1/auth/automation/authorize",
            "scopes": {
              "app.cancel": "Cancel durable application work.",
              "app.credentials.read": "Inspect credential status.",
              "app.credentials.write": "Change credential references.",
              "app.read": "Read application state.",
              "app.run": "Start durable application work.",
              "app.write": "Change reversible application state.",
              "manager.mutate": "Execute approved Manager plans.",
              "manager.read": "Inspect Manager state and plans.",
              "manager.runtime": "Control managed runtime services."
            },
            "tokenUrl": "/api/v1/auth/automation/token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "Pandrator API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/automation-clients": {
      "get": {
        "operationId": "listAutomationClients",
        "responses": {
          "200": {
            "description": "Automation clients"
          }
        }
      },
      "post": {
        "operationId": "registerAutomationClient",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutomationClientCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Registered"
          }
        }
      }
    },
...
    "/api/v1/auth/tokens": {
      "get": {
        "operationId": "listApiTokens",
        "responses": {
          "200": {
            "description": "Tokens"
          }
        }
      },
      "post": {
        "operationId": "createApiToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
...
    "/api/v1/jobs": {
      "get": {
        "operationId": "listJobs",
        "responses": {
          "200": {
            "description": "Jobs"
          }
        }
      },
      "post": {
        "operationId": "createJob",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Queued"
          }
        }
      }
    },
...
    "/api/v1/manager/plans": {
      "post": {
        "operationId": "createManagerPlan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManagerPlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Immutable manager plan"
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerToken": []
          },
          {
            "nativeOAuth": [
              "manager.read"
            ]
          }
        ]
      }
    },
...
    "/api/v1/providers": {
      "get": {
        "operationId": "listProviders",
        "responses": {
          "200": {
            "description": "Providers"
          }
        },
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerToken": []
          },
          {
            "nativeOAuth": [
              "app.read"
            ]
          }
        ]
      },
      "post": {
        "operationId": "createProvider",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
...
    "/api/v1/rvc/convert": {
      "post": {
        "operationId": "convertWithRvc",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RvcConvertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Queued"
          }
        }
      }
    },
...
    "/api/v1/session-bundles/import": {
      "post": {
        "operationId": "importSessionBundle",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BundleImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Queued"
          }
        }
      }
    },
...
    "/api/v1/training": {
      "get": {
        "operationId": "listTrainingRuns",
        "responses": {
          "200": {
            "description": "Training runs"
          }
        }
      },
      "post": {
        "operationId": "createTrainingRun",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Queued"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}